requirements.txt
Django==5.0.3
djangorestframework==3.15.1
drf-material==0.2.6
requests==2.31.0
pytest==8.1.1
python-dotenv==1.0.1
pytest-django==4.8.0
